requirements.txt
Django==4.2.7
Pillow>=10.0.0
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-extensions==3.2.3

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4

# Development and testing
coverage==7.3.2
django-debug-toolbar==4.2.0
factory-boy==3.3.0

# Security
django-ratelimit==4.1.0
django-cors-headers==4.3.1
